`render` callback for component does not return view.$
This is a breaking change.
Previously, calling `this.render()` in a module-for-component would
append the view and then return the view's `this.$()`. This changes
`render` to only render the component and append it, returning
`undefined`.
Test users can still use `this.$()` to get the component's jquery-wrapped element.

This avoids a possible issue in Ember >= 1.11.0-beta.5 where tagless
components (`tagName = ''`) will give a deprecation warning when their
`$()` method is called.