Remove isolatedContainer and use Ember.Application.buildRegistry.
The isolatedContainer was way to finicky and caused breakage any time Ember
or Ember Data decided to change what the default contents should be.

This replaces the `isolatedContainer` with the container built by
`Ember.Application.buildRegistry` (or `.buildContainer`) which has been available
since 1.0.0.  `.buildRegistry` does result in a container that is fully fleshed out
but it also has a resolver attached, so in the non-integration case we replace that
resolver with an empty function.