Ensure global view registry is aliased when view registry is used.
In Ember 1.12.0 and 1.12.1 the `-view-registry:main` was introduced as a
mechanism to track views without mutating the global `Ember.View.views`
registry (needed for fastboot work), unfortunately the
`Ember.EventDispatcher` was not updated to use the new view registry and
therefore any views added in a view registry other than
`Ember.View.views` will not be able to handle DOM events. The event
dispatcher was fixed in https://github.com/emberjs/ember.js/pull/11290.

This change ensures that `-view-registry:main` is the same as
`Ember.View.views` on Ember versions where `Ember.View.views` exists
(Ember < 1.13.0).