Add support for `integration: 'legacy'`
This semantics of the `integration` flag have changed in a
backwards-incompatible way since v0.4.3.  This corresponds to an ember upgrade
(using ember-cli) between v1.12 and v1.13.

In order to ease this upgrade path for ember-cli users, `integration: 'legacy'`
will set up a component test with v0.4.3 semantics.  Of course, newer tests
would still be encouraged to be written with `integration: true`.