Add `wait` helper.
This allows component integration tests that need to deal with async
interactions to have an easy way to model their tests.

Example test (assuming the component does some async in an action
triggered by the button click):

```javascript
import { moduleForComponent, test } from 'ember-test-helpers';
import waitForPendingRunLoops from 'ember-test-helpers/wait';

test('it works when async exists in an event/action', function(assert) {
  this.render('{{my-foo}}');

  this.$('button').click();

  return waitForPendingRunLoops()
    .then(() => {
      // assertions for after async behavior
    });
});
```

The guts of this `wait` helper are pulled from the `wait` acceptance
test helper in Ember itself. Eventually we may be able to make Ember
itself provide an implementation that we can share here and in the
acceptance testing world.