Ensure prior render is cleared when calling `this.render`.
Previously, calling `this.render` twice within the same test (without
calling `this.clearRender()` in between) would cause any components
created in the first render to never be destroyed (and therefore they
cannot cleanup any listeners they created).

This change essentially ensures that we call `clearRender` before each
`this.render` invocation so that if a component had been previously
instantiated it is cleared and reset.