Remove usage of `context` in `moduleForComponent`.
Setting `context` or `controller` on an `Ember.Component` was absolutely
not supposed to work (ever), but this library worked around that to
serve its own needs when `Ember.View` was deprecated and removed.

This PR makes `moduleForComponent` work properly against canary (soon to
be 2.7.0), by avoiding this archaic concept and simply tracking the
things `set` on the test context manually.