Fix regression related to DOM structure.
Ensures that `this.$()` from within a test represents the same structure
as prior versions.

Ultimately, I would prefer for `this.$()` to refer to the inner HTML
(basically what is rendered by the template provided to `this.render`),
but changing the underlying structure is definitely not good without
proper messaging (and likely major version bump).