Remove `TestModuleForIntegration`.
This was added in an early attempt to simplify / separate things for the
"grand testing unification" RFC, but ultimately was never rolled out to
folks.

At this point, it has become clear that this class based system for
managing test setup is not the way forward (see emberjs/rfcs#230 for
details) and this module (that is unused) has just become "dead weight"
maintenance wise.