Ensure ember-testing-container is reset properly.
The changes made in 7ad93d2c introduced a bug in the way that the
testing DOM is cleaned up between tests. The error was quite simple:
we were storing the `storedValue = container.outerHTML` before
starting the test then resetting `container.innerHTML = storedValue`.
This created a DOM node leak for each test processed.

This fixes the issue for both the new testing API's in 0.7 and the
older pre-existing system. It also adds some post-test checks to
ensure that the DOM is indeed what we expect after each test is
completed.