Refactor cleanup assertions to use global failure after test run.
Throwing during a `QUnit.testDone` callback causes the entire test suite
to abort (and does not actually fail the test in question anyways). This
changes things around so that we track these failures and throw them
after the test suite is completed, which is caught by QUnits global
`window.onerror` handler as a "global failure".