Do not shutdown the server while timers are still active.
Prior to this change, async timers were still pending (specifically the
nested ajax requests done in the `settled` integration tests) when we
shut down the pretender server. Then when the async which hits one of
the Pretender end points finally kicks off, the server was not present
and an error occurred.

This was previously being swallowed (because of horrid jQuery deferred
BS) but a prior fix made the error bubble up properly.