Remove tests requiring `click` only wait for subset of settledness.
These tests previously used a sync version of click that allowed
`settled` to be tested with _only_ waiting for some portions of
settledness. When the tests were converted to use the new public version
of `click` the test began failing (because `click` doesn't kick off
until the next tick of the event loop, so `settled` resolves
immediately).