Remove duplicated `wait` tests.
These tests are the exact same tests as exist in
`tests/integration/settled-test.js`, but with the `import wait from
'ember-test-helpers/wait'` instead of `import { settled } from
'@ember/test-helpers'`.  The two utility functions are _exactly_ the
same (`wait` is a reexport of `settled`), so there is no reason to test
both versions like this.

This fixes an annoying flakey test that keeps showing up in CI. The
corresponding test was already fixed in `settled-test.js` to handle
async properly, but the broken (and flakey) version still existed here
causing sporadic failures. Instead of fixing the issue and forcing all
future `settled-test.js` changes to have to remember to _also_ update
this duplicated file, I chose to delete it...