Refactor `setupRenderingContext` so that element is stable after setup.
Previously (with `moduleForComponent`) eager rendering was avoided
because it was not clear that rendering was _actually_ desired (and
therefore the extra work would be wasted in cases that were not testing
rendering). Unlike `moduleForComponent` we can be certain that
`setupRenderingContext` is only used in actual rendering contexts (its
in the name :P ), so we do not need to be concerned with the
performance.

This change is a compatible change (unless someone was _relying_ on
the fact that `this.element` was undefined, which seems super weird)
that ensures `this.element` is set immediately upon
`setupRenderingContext` being called (instead of lazily when the first
`render` is done). At that point `this.element` is essentially an empty
element (roughly akin to having only `{{outlet}}` in an `application`
template without an `index` template).

This unlocks the ability to test pending promises during rendering (via
things like `waitUntil`).