CI: Remove gh-pages auto-deployment.
It doesn't work at the moment due to an issue with emberjs/website
(which also deploys via gh-pages for the whole emberjs org).

For now (in order to get the docs "out there") we will add them as a
markdown to the repo that is updated when releases are prepared (e.g.
the same time as the changelog is generated).