Core: extract getRootElement functionality to util
The utility is private for now, but may be made public in the future
(needs more docs at the least).

The basic gist is that `getRootElement` should search relative to the
`rootElement` of the application (when `setApplication` was used), and
_not_ scoped to the first `.ember-view` _inside_ that element. This
ensures that elements added to the `rootElement` (e.g. via
`ember-wormhole` / `liquid-wormhole`) can also be interacted with.