Remove simple alias for `wait` -> `settled`.
The original goal of `settled` was to "rebrand" `wait` as `settled` without the
'baggage" (since `wait` had some legacy APIs that should not exist in
`settled`). Unfortunately, the code refactoring that I had originally done that
allowed them to be shared resulted in introducing bugs like those
reported.

This commit completely severs the implementation of `wait` from that of
`settled`, by leveraging the (IMHO very nice) layers of abstraction that we
have created since the original extraction.  Implementing `wait` with
`waitUntil` and `getSettledState` makes this quite trivial now.

tldr of changes:

* `isSettled` no longer has to support the legacy `wait` contract (this was
  undocumented private behavior _only_ kept around to support the old semantics
  of `wait`)
* `wait` is implemented as its own `waitUntil` with the same
  semantics it has had since the 0.6.x series