Adds a basic failing test for `settled`
As of Ember 3.2, `isSettled()` is no longer `true`, even immediately
after `await settled()`. This is because `waitUntil` uses an RSVP
Promise which resolves with a runloop. So when `waitUntil`'s promise
is about to resolve, things _are_ settled, but as soon as we actually
resolve, a new runloop is created which causes `isSettled()` to be
`false`.