Remove shared state work around.
The code added in an earlier commit was working around the underlying
issue. Essentially, using `getContext` in `hasPendingTransition` meant
that we were always using whatever the "last" context was and since
every test does not (and **should not**) call `setContext` we were
accidentally sharing an older tests context.

This change ensures we `setContext` up before these
`setupApplicationContext` tests (which is important), but we _also_
should add more changes to the testing harness in @ember/test-helpers
overall to ensure that each test internally cleans up the context (look
out for future PR's/commits for that)...