Only customize RSVP's async for Ember < 1.7.
As of Ember 1.7 and higher, this customization is done automatically.
Originally we intended to just leave this alone, and not update (since
these are already legacy testing systems) but apparently this is
actively breaking other fixes in the newer system if *any* older style
tests still exist in the test suite.

The main issue with this code on newer Ember versions is that it does
not properly call `adapter.asyncStart()` and `adapter.asyncEnd()` which
means that we no longer wait for the promise resolution in the test
harness itself. :sob: