Add pauseTestFor method.
Ensures that tests wait for the provided promise before being considered
"settled".

This function is expected to be used in application/addon code (aka
non-test code).  When used during a test run, it ensures that the next
`await settled()` check does not result until after the promise has
settled, when used outside of a test it is a simple "pass through"
function.