Fire blur on click/dbl-click/tap on non-focusable element
The root cause of the issue is that the blur event is not triggered on
the current active element in the __focus__ helper in case if the
clicked (double-clicked, tapped) element is not focusable.

If the clicked (double-clicked, tapped) element is focusable the
previous active element gets
blur event trough the __focus__ helper.

A call to __blur__ was added to the __focus__ helper for the case
when the clicked (double-clicked, tapped) element element
is not focusable.