Remove ember-data from test harness
We no longer have `moduleForModel` tests (:tada:), so we don't need
ember-data as a dependency any longer. This helps fix a number of issues
with our test suite running against canary builds, where the older
version of ember-data we were using was missing some required properties
of the `deprecate` invocation (but upgrading to newer ember-data wasn't
possible because of node engine support).