Bring back `{{#in-element}}` test in ember-classic
After reviewing further we determined that the issue that was causing
the failure with the `{{#in-element` setup is that `insertBefore` is
required in order to force "append" mode. Without `insertBefore` the
rendering engine will actually **clear** the contents of the target
element (which is waht causes the errors we were seeing).