Upgrade to Ember v4 types
Since this library was not generally using deprecated APIs, this simply
allows us to get the latest and greatest type definitions. The one
deprecated API that *is* in use is direct access to the `run`
namespace. However, that is already covered via a `hasEmberVersion`
check.

In several cases, introduce additional safety handling where TypeScript
can now identify failure modes (both because of improvements to the
compiler and because our tsconfig is now stricter). In others,
introduce hard errors for cases where TypeScript points out the need
for *some* variety of validation, but the failure scenario is one where
there is nothing actionable whatsoever.

Along the way, fix a number of lingering type errors, some of which
were pedantic but a number of which represented actual failure
conditions.