Publish types in `public-types` with `typesVersions`
Publish types to a `public-types` directory, in parallel to the
in-place Babel-powered emit for JS. In the future, we should probably
go ahead and remove the Babel emit and get this the rest of the way
into a v2 addon format with a single publish step. However, this
minimizes the delta introduced in this particular PR, and thus keeps
the risk lower.

This does *not* correctly publish types for `ember-test-helpers`.
However, there remains only one export from that package the
`has-ember-version` module and helper, which is also available from
`@ember/test-helpers`. Accordingly, we can simply ignore that (and plan
to go ahead and eventually remove it).