assert only what we legally can on Router
As with `RouterService`, the class is not exported (though its type
is), at least on the versions where we end up resolving it, since it is
not intended to be subclassed. Assert that it at least exists, and
introduce an unsafe cast and a `SAFETY` comment to explain the cast.