docs: publish updates to setupContext
`Commit: cba01bea` updated `setupContext`, renaming its first param from
`context` to `base` in the process. It did not regenerate out the docs
as part of the changeset, so building the docs results in this change
locally.

Let's fixup that commit here so that we can build the docs without
creating any local changes.