Make test context reusable during a test
In v2.0.0 @ember/test-helpers refactored to use
@ember/destroyable for teardown which also meant
that the application context was destroyed during
teardown, where previously it was only cleaned up
and the application was destroyed.

In our application tests we simulate a page
refresh by calling `teardownContext`` and
`setupContext`. The destroyable change broke this
functionality because the context is now marked
as destroyed.

This commit effectively reverts the change to use
@ember/destroyable and adds tests to ensure that
the context is not destroyed so that it can be
reset during a test.