[CLEANUP] remove manually annoted legacy code
These are slightly riskier in that you'd have to take the comment
author's word for it, as the proposed code was never execrised or
tested in the real world (unlike in #20798, where, if there were
no bugs in the branchgin, we should have been running the correct
branch all along and for a while now).

But I manually audited these spots and to my knowledge they seemed
safe.