Introduce `waitForFocus` DOM Helper
Convinient helper that resolves when a target receives focus. Useful for
verifying keyboard navigation handling and default focus.

Uses the pull based waitUntil helper to support the element not being in
the DOM when invoking the helper.

Alternatives without this helper is asserting `document.activeElement` is
the target. It usually work, but there are cases it focus may happne
async. The element isn't in view yet.